backend/agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
moviepy>=2.0.0
boto3>=1.38.0
